feat: keep track of expected WAL size
This patch keeps track of the segment size in the WAL file as written on
the Klio server. Doing that, it is possible to zero-pad partial WAL
files until they reach the requested segment length.

+++ internal/client/klioclient/common/errors.go
@@ -26,7 +26,7 @@ type IncompleteTransmissionError struct {
 
 	// WrittenBytes is the number of bytes that have successfully beings received
 	// by the server
-	WrittenBytes uint64
+	WrittenBytes int
 }
 
 // Error implements the error interface
